pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nnviz"
version = "0.1.0"
description = "Neural network region visualization tool"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "Pillow>=9.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
nnviz = "nnviz.__main__:main"

[tool.setuptools.packages.find]
include = ["nnviz*"]
